requirements.txt
contractions==0.1.72
matplotlib==3.4.3
nltk==3.6.5
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.1
scikit_learn==1.0.2
spacy==3.2.4
streamlit==1.8.1
tweepy==4.8.0
wordcloud==1.8.1
